Initial commit Snapshot of original group project
This is an initial commit representing the original state of the Git repository my group developed for our Multimedia Design & Programming class.

While I may want to make edits or improvements later, this snapshot reflects the final version we reached during the course.

Any code sizable piece of code or function that is not preceded by worked on by Hugo in a comment is from other groupmates. Speech To Text was my module. I didn't want to take credit for things I didn't do but worried about anonymity so I removed the portions of comments with credit.